First Commit
This is AXCS built using AHK v1

AXCS Voice Recognition is the main script that listens to the user.

AXCS Voice Commands are a library script that allows the user to execute specific commands i.e. "Open Facebook"

AXCS Source is a JSON document where the names of commands are saved. It may be redundant.

AXCS GUI is a buggy mess, that hopes to act as a GUI for the user so that they can freely create, edit, or manipulate their phrases and commands without having to use a notepad or IDE to make changes.